+++ src/Artsy/Router/buildClientApp.tsx
@@ -20,6 +20,8 @@ import { Boot } from "Artsy/Router/Boot"
 
 import { RouterConfig } from "./"
 
+import { RenderPending, RenderReady } from "./Utils/RenderStatus"
+
 interface Resolve {
   ClientApp: ComponentType<any>
 }
@@ -64,7 +66,12 @@ export function buildClientApp(config: RouterConfig): Promise<Resolve> {
         }),
       ]
       const resolver = new Resolver(relayEnvironment)
-      const render = createRender({})
+
+      const render = createRender({
+        renderPending: RenderPending,
+        renderReady: RenderReady,
+      })
+
       const Router = await createInitialFarceRouter({
         historyProtocol: getHistoryProtocol(),
         historyMiddlewares,